Extract authentication methods into separate service
