+++ backends/carp_webservices/lib/carp_auth/carp_auth_properties.dart
@@ -1,0 +1,63 @@
+/*
+ * Copyright 2018 Copenhagen Center for Health Technology (CACHET) at the
+ * Technical University of Denmark (DTU).
+ * Use of this source code is governed by a MIT-style license that can be
+ * found in the LICENSE file.
+ */
+part of 'carp_auth.dart';
+
+/// Represents a CARP web service app endpoint.
+class CarpAuthProperties {
+  /// The OAuth endpoint for this app.
+  final Uri authURL;
+
+  /// OAuth client id
+  final String clientId;
+
+  /// OAuth client secret
+  String? clientSecret;
+
+  /// Redirect URI for OAuth
+  final Uri redirectURI;
+
+  /// Redirect uri for OAuth after logout
+  /// If not specified, the [redirectURI] is used.
+  Uri? logoutRedirectURI;
+
+  /// Discovery URI for OAuth
+  final Uri discoveryURL;
+
+  /// The CARP study id for this app.
+  String? studyId;
+
+  /// The CARP study deployment id of this app.
+  String? studyDeploymentId;
+
+  /// Create a [CarpAuthProperties] which know how to access a CARP backend.
+  ///
+  /// [name], [uri], and [oauth] are required parameters in order to identify,
+  /// address, and authenticate this client.
+  ///
+  /// A [studyDeploymentId] and a [study] may be specified, if known at the
+  /// creation time.
+  CarpAuthProperties({
+    required this.authURL,
+    required this.clientId,
+    this.clientSecret,
+    required this.redirectURI,
+    required this.discoveryURL,
+    this.studyDeploymentId,
+    this.studyId,
+    this.logoutRedirectURI,
+  });
+
+  @override
+  int get hashCode => authURL.hashCode;
+
+  @override
+  bool operator ==(other) => authURL == other;
+
+  @override
+  String toString() =>
+      'CarpApp - auth URL: $authURL, studyDeploymentId: $studyDeploymentId, studyId: $studyId';
+}
